chore(tailwindcss-transformer): Update tests
