--- packages/tailwindcss-transformer/tests/render-prop.fixture.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export function RenderProp() {
-  return <Component render={() => <div className="flex" />} />
-}
