--- packages/tailwindcss-transformer/tests/slot.fixture.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export function Slot() {
-  return <Component slot={<div className="flex" />} />
-}
